moved components folder out of pages folder.
