+++ client/pages/schools/[dbn].js
@@ -1,8 +1,8 @@
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
-import Card from "../components/Card";
-import CardSquare from "../components/CardSquare";
-import SchoolButton from "../components/SchoolButton";
+import Card from "../../components/Card";
+import CardSquare from "../../components/CardSquare";
+import SchoolButton from "../../components/SchoolButton";
 
 function SchoolProfile() {
     const router = useRouter();
